requirements.txt
Flask==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
Flask-Login==0.6.3
bcrypt==4.0.1